Move feed models from model.commercial.feeds to commercial.model.feeds.
