+++ commercial/app/CommercialLifecycle.scala
@@ -1,11 +1,11 @@
 package commercial
 
-import app.LifecycleComponent
-import commercial.feeds._
-import common.LoggingField._
-import common._
-import metrics.MetricUploader
 import _root_.model.commercial.jobs.Industries
+import app.LifecycleComponent
+import commercial.model.feeds._
+import common.LoggingField._
+import common.{AkkaAsync, JobScheduler, Logging}
+import metrics.MetricUploader
 import play.api.inject.ApplicationLifecycle
 
 import scala.concurrent.{ExecutionContext, Future}
