+++ commercial/app/model/feeds/FeedFetcher.scala
@@ -1,4 +1,4 @@
-package commercial.feeds
+package commercial.model.feeds
 
 import java.lang.System.currentTimeMillis
