+++ commercial/app/model/feeds/FeedMetaData.scala
@@ -1,9 +1,8 @@
-package commercial.feeds
+package commercial.model.feeds
 
-import commercial.feeds.ResponseEncoding.utf8
+import commercial.model.feeds.ResponseEncoding.utf8
 import conf.switches.{Switch, Switches}
 import model.commercial.soulmates.SoulmatesAgent
-import org.joda.time.{DateTime, DateTimeZone}
 
 import scala.concurrent.duration.{Duration, _}
