+++ commercial/app/model/merchandise/events/LiveEventAgent.scala
@@ -2,13 +2,12 @@ package model.commercial.events
 
 import java.lang.System._
 
-import commercial.feeds._
+import commercial.model.feeds._
 import common.{AkkaAgent, ExecutionContexts, Logging}
 import conf.Configuration
 import model.commercial.LiveEvent
 import play.api.libs.json.JsValue
 import play.api.libs.ws.{WSClient, WSResponse}
-import commercial.model.readsSeq
 
 import scala.concurrent.Future
 import scala.concurrent.duration._
