+++ commercial/app/model/merchandise/soulmates/SoulmatesFeed.scala
@@ -2,10 +2,9 @@ package model.commercial.soulmates
 
 import java.lang.System._
 
-import commercial.feeds.{FeedMetaData, MissingFeedException, ParsedFeed, SwitchOffException}
+import commercial.model.feeds.{FeedMetaData, MissingFeedException, ParsedFeed, SwitchOffException}
 import common.{ExecutionContexts, Logging}
 import model.commercial.Member
-import commercial.model.readsSeq
 import play.api.libs.json.{JsValue, Json}
 
 import scala.concurrent.Future
