+++ commercial/app/model/merchandise/travel/TravelOffersApi.scala
@@ -2,7 +2,7 @@ package model.commercial.travel
 
 import java.lang.System.currentTimeMillis
 
-import commercial.feeds.{FeedMetaData, MissingFeedException, ParsedFeed, SwitchOffException}
+import commercial.model.feeds.{FeedMetaData, MissingFeedException, ParsedFeed, SwitchOffException}
 import common.{ExecutionContexts, Logging}
 import model.commercial.TravelOffer
 import org.joda.time.format.DateTimeFormat
